refactor: move files, add example hooks and constants
