+++ packages/react-apollo/src/shared/constants.ts
@@ -1,1 +1,5 @@
 export const IS_DEV = import.meta.env.DEV;
+
+export enum LocalStorageKeys {
+  StoredId = 'storedId',
+}
